fix(docs): keep changelog parser server-only
