+++ docs/lib/parse-changelog.ts
@@ -420,19 +420,3 @@ export async function parseChangelog(rootDir: string): Promise<ChangelogEntry[]>
   const sources = await loadChangelogSources(rootDir);
   return parseChangelogSources(sources);
 }
-
-export function getEntryPreviewHtml(entry: Pick<ChangelogEntry, "contentBlocks">): string {
-  const block = entry.contentBlocks.find(
-    (contentBlock): contentBlock is Extract<ChangelogContentBlock, { type: "markdown" }> =>
-      contentBlock.type === "markdown",
-  );
-
-  return block?.html ?? "";
-}
-
-export function getEntrySearchText(entry: Pick<ChangelogEntry, "contentBlocks">): string {
-  return entry.contentBlocks
-    .map((block) => (block.type === "markdown" ? block.plainText : block.code))
-    .join(" ")
-    .trim();
-}
